==> 00- Changing_compression.ipynb <==
## Importing Libraries

from zipfile import ZipFile
import os
import re
import gzip

## finding files with zip extensions in './data/' directory

data_folder = "./data/"

# Loading the list of file in './data/' directory
pattern = r'(.*)\.zip'
files_n = [[par_n+"/",re.match(pattern, file)[1]] for par_n, dir_n, file_n in os.walk(data_folder) 
           for file in file_n if re.match(pattern, file)!=None]

# Changing the compression to gzip
for path, file in files_n:
    # Extracting the zip file
    with ZipFile(path+file+'.zip', 'r') as zipf:
        zipf.extractall(path)
    # Compressing the extracted text file using gzip compression
    with open(path+file+'.txt', 'rb') as f_in, gzip.open(path+file+'.txt.gz', 'wb') as f_out:
        f_out.writelines(f_in)
    # Removing the extracted text file
    os.remove(path+file+'.txt')

## Uploading gzipped files

# For uploading the gzipped files, run the following in command line 
# (assuming all the gzip files are directly inside ./data/ directory on your local machine)
gsutil -m cp -r /data/*.gz gs://my-bucket/directory
