==> 00- Creating a Sample Data.ipynb <==
import pandas as pd
import numpy as np
pd.set_option('display.max_columns', 100)

## 1. Creating a Sample of "tls201" table

data_folder = './data/data_PATSTAT_Global_2018_Autumn_01/'

df = pd.read_csv(data_folder+'tls201_part01.txt.gz', compression='gzip', low_memory=False, nrows=100000)
df.head()

df.columns

df.to_csv(data_folder+'tls201_part01.txt', encoding='utf-8', index=False)

## 2. Creating a sample of "tls209" table: 

data_folder = './data/data_PATSTAT_Global_2018_Autumn/'

df = pd.read_csv(data_folder+'tls209_part01.txt.gz', compression='gzip', low_memory=False, nrows=100000)
df.head()

df.to_csv(data_folder+'tls209_part01.txt', encoding='utf-8', index=False)

## 3. Creating a sample of "tls204" table: Priority patents

data_folder = './data/data_PATSTAT_Global_2018_Autumn/'

df = pd.read_csv(data_folder+'tls204_part01.txt.gz', compression='gzip', low_memory=False, nrows=100000)
df.head()

df.to_csv(data_folder+'tls204_part01.txt', encoding='utf-8', index=False)

## 4. Creating a sample of "tls207": 
## Lists the correspondence between patent application and inventors (applicants)

data_folder = './data/data_PATSTAT_Global_2018_Autumn/'

df = pd.read_csv(data_folder+'tls207_part01.txt.gz', compression='gzip', low_memory=False, nrows=100000)
df.head()

df.to_csv(data_folder+'tls207_part01.txt', encoding='utf-8', index=False)

## 5. Creating a sample of "tls206": 
## Provides details on names and addresses of applicants

data_folder = './data/data_PATSTAT_Global_2018_Autumn/'

df = pd.read_csv(data_folder+'tls206_part01.txt.gz', compression='gzip', low_memory=False, nrows=100000)
df.head()

df.to_csv(data_folder+'tls206_part01.txt', encoding='utf-8', index=False)

## 6. Creating a sample of "tls211": 
## Information on the patent offices of destination (publication authorities) of all INPADOC family members and excludes the PCT publication authority (WO)

data_folder = './data/data_PATSTAT_Global_2018_Autumn/'

df = pd.read_csv(data_folder+'tls211_part01.txt.gz', compression='gzip', low_memory=False, nrows=100000)
df.head()

type(df.iloc[0,4])

df.to_csv(data_folder+'tls211_part01.txt', encoding='utf-8', index=False)

## 6. Creating a sample of "tls212": Citations

data_folder = './data/data_PATSTAT_Global_2018_Autumn/'

df = pd.read_csv(data_folder+'tls212_part01.txt.gz', compression='gzip', low_memory=False, nrows=100000)
df.head()

## 7. Creating a sample of "tls203": Application Abstract

data_folder = '../../patstat_2018_backup/openPatstat/data/data_PATSTAT_Global_2018_Autumn/'

df = pd.read_csv(data_folder+'tls203_part01.txt.gz', compression='gzip', low_memory=False, nrows=100000)
df.head()

## 8. Creating a sample of "tls205": Technical Relation

data_folder = '../../patstat_2018_backup/openPatstat/data/data_PATSTAT_Global_2018_Autumn/'

df = pd.read_csv(data_folder+'tls205_part01.txt.gz', compression='gzip', low_memory=False, nrows=100000)
df.head()

## 9. Creating a sample of "tls210": National Classification

data_folder = '../../patstat_2018_backup/openPatstat/data/data_PATSTAT_Global_2018_Autumn/'

df = pd.read_csv(data_folder+'tls210_part01.txt.gz', compression='gzip', low_memory=False, nrows=100000)
df.head()

## 10. Creating a sample of "tls214": Non patent literature publication

data_folder = '../../patstat_2018_backup/openPatstat/data/data_PATSTAT_Global_2018_Autumn/'

df = pd.read_csv(data_folder+'tls214_part01.txt.gz', compression='gzip', low_memory=False, nrows=100000)
df.head()

df.columns

## 11. Creating a sample of "tls215": Citation Category

data_folder = '../../patstat_2018_backup/openPatstat/data/data_PATSTAT_Global_2018_Autumn/'

df = pd.read_csv(data_folder+'tls215_part01.txt.gz', compression='gzip', low_memory=False, nrows=100000)
df.head()

## 12. Creating a sample of "tls216": Application Continuation

data_folder = '../../patstat_2018_backup/openPatstat/data/data_PATSTAT_Global_2018_Autumn/'

df = pd.read_csv(data_folder+'tls216_part01.txt.gz', compression='gzip', low_memory=False, nrows=100000)
df.head()

## 13. Creating a sample of "tls222": Japanese Classification

data_folder = '../../patstat_2018_backup/openPatstat/data/data_PATSTAT_Global_2018_Autumn/'

df = pd.read_csv(data_folder+'tls222_part01.txt.gz', compression='gzip', low_memory=False, nrows=100000)
df.head()

## 14. Creating a sample of "tls223": US Classification

data_folder = '../../patstat_2018_backup/openPatstat/data/data_PATSTAT_Global_2018_Autumn/'

df = pd.read_csv(data_folder+'tls223_part01.txt.gz', compression='gzip', low_memory=False, nrows=100000)
df.head()

## 15. Creating a sample of "tls226": Unmodified Person Data

data_folder = '../../patstat_2018_backup/openPatstat/data/data_PATSTAT_Global_2018_Autumn/'

df = pd.read_csv(data_folder+'tls226_part01.txt.gz', compression='gzip', low_memory=False, nrows=100000)
df.head()

len(df.columns)

## 16. Creating a sample of "tls227": Link Between Person and Publication

data_folder = '../../patstat_2018_backup/openPatstat/data/data_PATSTAT_Global_2018_Autumn/'

df = pd.read_csv(data_folder+'tls227_part01.txt.gz', compression='gzip', low_memory=False, nrows=100000)
df.head()

## 17. Creating a sample of "tls229": Industry Classification

data_folder = '../../patstat_2018_backup/openPatstat/data/data_PATSTAT_Global_2018_Autumn/'

df = pd.read_csv(data_folder+'tls229_part01.txt.gz', compression='gzip', low_memory=False, nrows=100000)
df.head()

## 18. Creating a sample of "tls230": Classification by Technical Field

data_folder = '../../patstat_2018_backup/openPatstat/data/data_PATSTAT_Global_2018_Autumn/'

df = pd.read_csv(data_folder+'tls230_part01.txt.gz', compression='gzip', low_memory=False, nrows=100000)
df.head()

## 19. Creating a sample of "tls231": Legal Event

data_folder = '../../patstat_2018_backup/openPatstat/data/data_PATSTAT_Global_2018_Autumn/'

df_1 = pd.read_csv(data_folder+'tls231_part01.txt.gz', compression='gzip', low_memory=False, nrows=100000)
df_1.head()

pf.merge(df_1, df_, on=['event_code'])

df[df.event_code=='A1B']

## 20. Creating a sample of "tls801": Reference table of country codes

data_folder = '../../patstat_2018_backup/openPatstat/data/data_PATSTAT_Global_2018_Autumn/'

df = pd.read_csv(data_folder+'tls801_part01.txt.gz', compression='gzip', low_memory=False, nrows=100000)
df.head()

## 21. Creating a sample of "tls803": Reference table of legal event codes

data_folder = '../../patstat_2018_backup/openPatstat/data/data_PATSTAT_Global_2018_Autumn/'

df = pd.read_csv(data_folder+'tls803_part01.txt.gz', compression='gzip', low_memory=False, nrows=100000)
df.head()

df[df.event_category_code=='F'].event_code.head(10)

## 22. Creating a sample of "tls901": Mapping between technology fields and IPC

data_folder = '../../patstat_2018_backup/openPatstat/data/data_PATSTAT_Global_2018_Autumn/'

df = pd.read_csv(data_folder+'tls901_part01.txt.gz', compression='gzip', low_memory=False, nrows=100000)
df.head()

## 23. Creating a sample of "tls902": Mapping between IPC and industrial sectors

data_folder = '../../patstat_2018_backup/openPatstat/data/data_PATSTAT_Global_2018_Autumn/'

df = pd.read_csv(data_folder+'tls902_part01.txt.gz', compression='gzip', low_memory=False, nrows=100000)
df.head()

## 24. Creating a sample of "tls904": NUTS regional codes

data_folder = '../../patstat_2018_backup/openPatstat/data/data_PATSTAT_Global_2018_Autumn/'

df = pd.read_csv(data_folder+'tls904_part01.txt.gz', compression='gzip', low_memory=False, nrows=100000)
df.head()

## 25. Creating a sample of "tls906": Person

data_folder = '../../patstat_2018_backup/openPatstat/data/data_PATSTAT_Global_2018_Autumn/'

df = pd.read_csv(data_folder+'tls906_part01.txt.gz', compression='gzip', low_memory=False, nrows=100000)
df.head()
