==> 00- Creating Big Query Tables.ipynb <==
from google.cloud import bigquery

from open_patstat.utils.gcp import create_table, load_gcs_file, delete_table
from open_patstat.utils.schema import Schema

client = bigquery.Client()

job_config = bigquery.LoadJobConfig()
job_config.skip_leading_rows = 1
job_config.max_bad_records = 10
job_config.source_format = bigquery.SourceFormat.CSV
dataset_ref = client.dataset('patstat')

# 1. Loading Tables

## 1.1. Table "tls201" (main table)

create_table(client,
             dataset_id='patstat',
             table_id='tls201',
             schema=Schema().tls201)

table_ref = dataset_ref.table('tls201_s')
job_config.schema = Schema().tls201
load_gcs_file(client, 'gs://patstat_2018g/data_PATSTAT_Global_2018_Autumn_01/tls201_*.gz', 
              table_ref, job_config)
#load_gcs_file(client, 'gs://patstat_2018g/data_PATSTAT_Global_2018_Autumn_01/tls201_part01.txt', 
#              table_ref, job_config)

## 1.2. Table "tls209_appln_ipc" (IPC codes assigned to each application)

# Creating main table
create_table(client,
             dataset_id='patstat',
             table_id='tls209',
             schema=Schema().tls209)

# Adding files to the main table from GCP bucket
table_ref = dataset_ref.table('tls209')
job_config.schema = Schema().tls209
load_gcs_file(client, 'gs://patstat_2018g/data_PATSTAT_Global_2018_Autumn/tls209_*.gz', 
              table_ref, job_config)

# Creating Sample dataset
create_table(client,
             dataset_id='patstat',
             table_id='tls209_s',
             schema=Schema().tls209)

# Adding files to the sample table from GCP bucket
table_ref = dataset_ref.table('tls209_s')
job_config.schema = Schema().tls209
load_gcs_file(client, 'gs://patstat_2018g/data_PATSTAT_Global_2018_Autumn/tls209_part01.txt', 
              table_ref, job_config)

## 1.3. Table "tls204" (the priority status of patents)

create_table(client,
             dataset_id='patstat',
             table_id='tls204',
             schema=Schema().tls204)

# Adding files to the main table from GCP bucket
table_ref = dataset_ref.table('tls204')
job_config.schema = Schema().tls204
load_gcs_file(client, 'gs://patstat_2018g/data_PATSTAT_Global_2018_Autumn/tls204_*.gz', 
              table_ref, job_config)

# Creating Sample dataset
create_table(client,
             dataset_id='patstat',
             table_id='tls204_s',
             schema=Schema().tls204)

# Adding files to the sample table from GCP bucket
table_ref = dataset_ref.table('tls204_s')
job_config.schema = Schema().tls204
load_gcs_file(client, 'gs://patstat_2018g/data_PATSTAT_Global_2018_Autumn/tls204_part01.txt', 
              table_ref, job_config)

## 1.4. Table "tls207": the correspondence between patent application and inventors

# Creating main table
create_table(client,
             dataset_id='patstat',
             table_id='tls207',
             schema=Schema().tls207)

# Adding files to the main table from GCP bucket
table_ref = dataset_ref.table('tls207')
job_config.schema = Schema().tls207
load_gcs_file(client, 'gs://patstat_2018g/data_PATSTAT_Global_2018_Autumn/tls207_*.gz', 
              table_ref, job_config)

# Creating Sample dataset
create_table(client,
             dataset_id='patstat',
             table_id='tls207_s',
             schema=Schema().tls207)

# Adding files to the sample table from GCP bucket
table_ref = dataset_ref.table('tls207_s')
job_config.schema = Schema().tls207
load_gcs_file(client, 'gs://patstat_2018g/data_PATSTAT_Global_2018_Autumn/tls207_part01.txt', 
              table_ref, job_config)

## 1.5. Table "tls206": Details on names and addresses of applicants

# Creating main table
create_table(client,
             dataset_id='patstat',
             table_id='tls206',
             schema=Schema().tls206)

# Adding files to the main table from GCP bucket
table_ref = dataset_ref.table('tls206')
job_config.schema = Schema().tls206
load_gcs_file(client, 'gs://patstat_2018g/data_PATSTAT_Global_2018_Autumn/tls206_*.gz', 
              table_ref, job_config)

# Creating Sample dataset
create_table(client,
             dataset_id='patstat',
             table_id='tls206_s',
             schema=Schema().tls206)

# Adding files to the sample table from GCP bucket
table_ref = dataset_ref.table('tls206_s')
job_config.schema = Schema().tls206
load_gcs_file(client, 'gs://patstat_2018g/data_PATSTAT_Global_2018_Autumn/tls206_part01.txt', 
              table_ref, job_config)

## 1.6. Table "tls211": 
## Information on the patent offices of destination (publication authorities) of all INPADOC family members and excludes the PCT publication authority (WO)

# Creating main table
create_table(client,
             dataset_id='patstat',
             table_id='tls211',
             schema=Schema().tls211)

# Adding files to the main table from GCP bucket
table_ref = dataset_ref.table('tls211')
job_config.schema = Schema().tls211

load_job = client.load_table_from_uri(
        source_uris='gs://patstat_2018g/data_PATSTAT_Global_2018_Autumn/tls211_*.gz',
        destination=table_ref,
        # job_id=job_id,
        job_id_prefix='lgs-',
        job_config=job_config,
    )

load_job.result()

# Creating Sample dataset
create_table(client,
             dataset_id='patstat',
             table_id='tls211_s',
             schema=Schema().tls211)

# Creating Sample dataset
create_table(client,
             dataset_id='patstat',
             table_id='tls211_s',
             schema=Schema().tls211)

# Adding files to the sample table from GCP bucket
table_ref = dataset_ref.table('tls211_s')
job_config.schema = Schema().tls211
load_gcs_file(client, 'gs://patstat_2018g/data_PATSTAT_Global_2018_Autumn/tls211_part01.txt', 
              table_ref, job_config)

## 1.7. Table "tls212": Citations

# Creating main table
create_table(client,
             dataset_id='patstat',
             table_id='tls212',
             schema=Schema().tls212)

# Adding files to the sample table from GCP bucket
table_ref = dataset_ref.table('tls212')
job_config.schema = Schema().tls212
load_gcs_file(client, 'gs://patstat_2018g/data_PATSTAT_Global_2018_Autumn/tls212_*.gz', 
              table_ref, job_config)

## 1.8. Checking to see if the data is loaded correctly on the tables

# Check table creation and displays meta data
# Run in command line
bq show my-dataset:patstat.tls201
