==> 00- Google SDK.ipynb <==
## 1. Copying "tls209" table from Local to Bucket (multi-threat and all subdirectories)

gsutil -m cp -r /home/reza/patstat2018/openPatstat/data/data_PATSTAT_Global_2018_Autumn/tls209_*.txt.gz 
                gs://patstat_2018g/data_PATSTAT_Global_2018_Autumn/

## Which Datasets are uploaded?

tls201
tls204
tls209
tls207

# Linux Commands:

## 1. Moving objects

mv -v Downloads/familyInformation.csv.zip patstat2018/family_patents/data/

## 2. Creating Sub directories

mkdir -p patstat2018/family_patent/data/

# setting folder permissions while making the directory
mkdir -m777 folder_name

# 3. Renaming a folder or directory

mv data_01.txt data_02.txt

# 4. changing the permissions of a file or folder

The `chmode` let you specify three numbers which set the permissions:

- Read = 4
- Write = 2
- Execute = 1

     And the `first number` is owner's access, `second` is group access, and `third number` is anybody else

# Owner: full access, group: read and execute, anyone else: no permissions
chmod 750 test
