==> 01- Big Query Script.ipynb <==
## Importing

from google.cloud import bigquery

from open_patstat.utils.gcp import create_table, load_gcs_file, delete_table
from open_patstat.utils.schema import Schema

## Initializing the Client anf Job Config

# Before running this line, make sure that you have defined the environment variable...
# ..."GOOGLE_APPLICATION_CREDENTIALS" which points to the JSON file containing authentication key
client = bigquery.Client()

# Initializing the Job_config
job_config = bigquery.LoadJobConfig()
job_config.skip_leading_rows = 1
job_config.max_bad_records = 10
job_config.source_format = bigquery.SourceFormat.CSV
dataset_ref = client.dataset('patstat')

## Creating and Adding Tables

########## Creating and Loading tables ##########
#################################################

# Tables list to be loaded
tables_list = ['tls212']

# Google Bucket directory address, which contains all data files
gs_add = 'gs://patstat_2018g/data_PATSTAT_Global_2018_Autumn/'

# Loading the tables in the list
for table in tables_list:
    # Creating the table
    create_table(client,
             dataset_id='patstat',
             table_id=table,
             schema=getattr(Schema(),table))
    # Adding files to the table from GCP bucket
    table_ref = dataset_ref.table(table)
    job_config.schema = getattr(Schema(),table)
    load_job = client.load_table_from_uri(
        source_uris=gs_add+table+'_*.gz',
        destination=table_ref,
        # job_id=job_id,
        job_id_prefix='lgs-',
        job_config=job_config,
    )
    load_job.result()

## 2. Creating "familyInfo" table

client = bigquery.Client()

# Initializing the Job_config
job_config = bigquery.LoadJobConfig()
job_config.skip_leading_rows = 1
job_config.max_bad_records = 10
job_config.source_format = bigquery.SourceFormat.CSV
dataset_ref = client.dataset('results_docdb')

########## Creating and Loading tables ##########
#################################################

# Tables list to be loaded
tables_list = ['familyInformation']

# Google Bucket directory address, which contains all data files
gs_add = 'gs://family_Information/'

# Loading the tables in the list
for table in tables_list:
    # Creating the table
    create_table(client,
             dataset_id='patstat',
             table_id=table,
             schema=getattr(Schema(),table))
    # Adding files to the table from GCP bucket
    table_ref = dataset_ref.table(table)
    job_config.schema = getattr(Schema(),table)
    load_job = client.load_table_from_uri(
        source_uris=gs_add+table+'_*.gz',
        destination=table_ref,
        # job_id=job_id,
        job_id_prefix='lgs-',
        job_config=job_config,
    )
    load_job.result()

## Adding Mapping table for mapping USPTO ID to PatStat ID

client = bigquery.Client()
# Initializing the Job_config
job_config = bigquery.LoadJobConfig()
job_config.skip_leading_rows = 1
job_config.max_bad_records = 10
job_config.source_format = bigquery.SourceFormat.CSV
dataset_ref = client.dataset('patstat')

########## Creating and Loading tables ##########
#################################################

# Tables list to be loaded
tables_list = ['mapping']

# Google Bucket directory address, which contains all data files
gs_add = 'gs://uspto-data/patentsView/'

# Loading the tables in the list
for table in tables_list:
    # Creating the table
    create_table(client,
             dataset_id='patstat',
             table_id=table,
             schema=getattr(Schema(),table))
    # Adding files to the table from GCP bucket
    table_ref = dataset_ref.table(table)
    job_config.schema = getattr(Schema(),table)
    load_job = client.load_table_from_uri(
        source_uris=gs_add+table+'*.gz',
        destination=table_ref,
        # job_id=job_id,
        job_id_prefix='lgs-',
        job_config=job_config,
    )
    load_job.result()

load_job.errors

import pandas as pd
data_folder = './data/'
