==> 02- Queries - Using DOCDB family ID.ipynb <==
import pandas as pd
from google.cloud import bigquery

# 1. Copying USPatents table

# from google.cloud import bigquery
client = bigquery.Client()
# Defining source_table
source_dataset = client.dataset('results', project='usptobias')
source_table_ref = source_dataset.table('USPatents')
# Defining target table
target_dataset = client.dataset('results_docdb', project='usptobias')
dest_table_ref = target_dataset.table('USPatents')
# Initializing the Job
job = client.copy_table(
    source_table_ref,
    dest_table_ref,
    # Location must match that of the source and destination tables.
    location='US')  # API request

# Running the job
job.result()  # Waits for job to complete.

## 1.2. Deleting the table

# from google.cloud import bigquery
# client = bigquery.Client()
dataset_id = 'results_docdb'
table_id = 'USPatents'

table_ref = client.dataset(dataset_id).table(table_id)
#client.delete_table(table_ref)  # API request

print("Table {}:{} deleted.".format(dataset_id, table_id))

# 2. USOnDocdb: USPatents Join on Docdb with tls201

client = bigquery.Client()

## 2.1. First Query: Generating merged table

# Creating Job Config
job_config = bigquery.QueryJobConfig()
#job_config.dry_run = True
job_config.use_query_cache = False
# Set configuration.query.writeDisposition
job_config.write_disposition = 'WRITE_TRUNCATE'

# Set the destination table
dataset_id = 'results_docdb'
table_id = 'USOnDocdb'
table_ref = client.dataset(dataset_id).table(table_id)
job_config.destination = table_ref

sql = """
    SELECT DISTINCT
            t1.appln_id AS appln_id_1,
            t1.appln_nr AS appln_nr_1,
            t1.appln_auth AS appln_auth_1,
            t1.appln_filing_year AS appln_filing_year_1,
            t1.appln_kind AS appln_kind_1,
            t2.appln_id AS appln_id_2,
            t2.appln_nr AS appln_nr_2,
            t2.appln_auth AS appln_auth_2,
            t2.appln_filing_year AS appln_filing_year_2,
            t2.appln_kind AS appln_kind_2,
            t1.docdb_family_id AS docdb_family_id
    FROM
            `usptobias.results_docdb.USPatents` AS t1
        LEFT JOIN
            `usptobias.patstat.tls201` AS t2
        ON
            t1.docdb_family_id=t2.docdb_family_id
    WHERE
            t1.appln_id<>t2.appln_id
     """
# Defining the query
query_job = client.query(
    sql,
    # Location must match that of the dataset(s) referenced in the query.
    location='US',
    job_config=job_config)

#            AND t2.appln_kind = 'A '
#            AND t2.appln_filing_year BETWEEN 2001 AND 2018

query_job.result()

## 2.2. Getting Information of USOnInpadoc table

# from google.cloud import bigquery
# client = bigquery.Client()
dataset_id = 'results_docdb'
table_id = 'USOnDocdb'

dataset_ref = client.dataset(dataset_id)
table_ref = dataset_ref.table(table_id)
table = client.get_table(table_ref)  # API Request

# View table properties
print(table.schema)
print(table.description)
print("\nThere are {:,} number of records in total!".format(table.num_rows))

# from google.cloud import bigquery
# client = bigquery.Client()
dataset_id = 'results_docdb'
table_id = 'USOnDocdb'

dataset_ref = client.dataset(dataset_id)
table_ref = dataset_ref.table(table_id)
table = client.get_table(table_ref)  # API Request

# View table properties
print(table.schema)
print(table.description)
print("\nThere are {:,} number of records in total!".format(table.num_rows))

## 2.3. Second Query: Counting Number of Pairs

# Creating Job Config
job_config = bigquery.QueryJobConfig()
#job_config.dry_run = True
job_config.use_query_cache = False
# Set configuration.query.createDisposition
job_config.create_disposition = 'CREATE_IF_NEEDED'
# Set configuration.query.writeDisposition
job_config.write_disposition = 'WRITE_TRUNCATE'

# Set the destination table
dataset_id = 'results_docdb'
table_id = 'USOnDocdb2'
table_ref = client.dataset(dataset_id).table(table_id)
job_config.destination = table_ref

# Number of pairs
sql = """
    SELECT
            t1.*,
            t2.docdb_nb_pairs
    FROM
            `usptobias.results_docdb.USOnDocdb` AS t1
        LEFT JOIN
            (
            SELECT
                    docdb_family_id,
                    appln_auth_2,
                    COUNT(*) AS docdb_nb_pairs
            FROM
                    `usptobias.results_docdb.USOnDocdb`
            GROUP BY
                    docdb_family_id, appln_auth_2
            ) AS t2
        ON
            (t1.docdb_family_id = t2.docdb_family_id) AND (t1.appln_auth_2 = t2.appln_auth_2)
     """
# Defining the query
query_job = client.query(
    sql,
    # Location must match that of the dataset(s) referenced in the query.
    location='US',
    job_config=job_config)

query_job.result()

## 2.4. Getting the information of "USOnInpadoc2" table

# from google.cloud import bigquery
# client = bigquery.Client()
dataset_id = 'results_docdb'
table_id = 'USOnDocdb2'

dataset_ref = client.dataset(dataset_id)
table_ref = dataset_ref.table(table_id)
table = client.get_table(table_ref)  # API Request

# View table properties
print(table.schema)
print(table.description)
print("\nThere are {:,} number of records in total!".format(table.num_rows))

# from google.cloud import bigquery
# client = bigquery.Client()
dataset_id = 'results_docdb'
table_id = 'USOnDocdb2'

dataset_ref = client.dataset(dataset_id)
table_ref = dataset_ref.table(table_id)
table = client.get_table(table_ref)  # API Request

# View table properties
print(table.schema)
print(table.description)
print("\nThere are {:,} number of records in total!".format(table.num_rows))

## 2.5. Loading 100 rows from the table

# from google.cloud import bigquery
# client = bigquery.Client()
dataset_id = 'results_docdb'
table_id = 'USOnDocdb2'

dataset_ref = client.dataset(dataset_id, project="usptobias")
table_ref = dataset_ref.table(table_id)
table = client.get_table(table_ref)  # API call

## Getting the first 100 lines of codes
# Load the first 10 rows
rows = client.list_rows(table, max_results=10000)

#pd.DataFrame(rows)
df = rows.to_dataframe()

df.head(7)

# 2.6. Deleting the table

# from google.cloud import bigquery
# client = bigquery.Client()
dataset_id = 'results_docdb'
table_id = 'USOnDocdb2'

table_ref = client.dataset(dataset_id).table(table_id)
#client.delete_table(table_ref)  # API request

print("Table {}:{} deleted.".format(dataset_id, table_id))

# 3. Exctrating exact_twins

client = bigquery.Client()

## 3.1. Query to generate the table

job_config = bigquery.QueryJobConfig()
job_config.user_query_cache = False
# Set configuration.query.writeDisposition
job_config.write_disposition = 'WRITE_TRUNCATE'

# Set destination table
dataset_id = 'results_docdb'
table_id = 'USExactTwins'
table_ref = client.dataset(dataset_id).table(table_id)
job_config.destination = table_ref

sql= """
    SELECT *
    FROM (
        SELECT
                *,
                (
                CASE 
                        WHEN docdb_nb_pairs=1 THEN 1
                        ELSE 0
                END
                ) AS exact_twins
        FROM
                `usptobias.results_docdb.USOnDocdb2`
         )
    WHERE
            appln_auth_2<>'US'
    """

# Defining the query job
query_job = client.query(sql, location='US', job_config=job_config)

query_job.result()

## 3.2. Getting the table information

dataset_id = 'results_docdb'
table_id = 'USExactTwins'
# Getting the table's information
table_ref = client.dataset(dataset_id).table(table_id)
table = client.get_table(table_ref)

# Print tables information
print(table.schema)
print(table.description)
print("\nThere are {:,} number of records in total!".format(table.num_rows))

## 3.3. Loading the first 1000 rows of the table

dataset_id = 'results_docdb'
table_id = 'USExactTwins'
# Getting the table object
table_ref = client.dataset(dataset_id).table(table_id)
table = client.get_table(table_ref)

# Load the first 10000 rows into RowIterator object
rows = client.list_rows(table, max_results=10000)

# Loading the RowIterator object into Pandas dataframe
df = rows.to_dataframe()

df.head(7)

# 3.4. Deleting the table

# from google.cloud import bigquery
# client = bigquery.Client()
dataset_id = 'results_docdb'
table_id = 'USExactTwins'

table_ref = client.dataset(dataset_id).table(table_id)
#client.delete_table(table_ref)  # API request

print("Table {}:{} deleted.".format(dataset_id, table_id))

# 4. Adding Priority filing information

# 4.1. Query

client = bigquery.Client()

job_config = bigquery.QueryJobConfig()
job_config.user_query_cache = False
# Set configuration.query.writeDisposition
job_config.write_disposition = 'WRITE_TRUNCATE'

# Set destination table
dataset_id = 'results_docdb'
table_id = 'backbone_0'
table_ref = client.dataset(dataset_id).table(table_id)
job_config.destination = table_ref

sql= """
    SELECT
            *,
            (
            CASE 
                    WHEN appln_id_1 NOT IN(
                                    SELECT DISTINCT appln_id
                                    FROM `usptobias.patstat.tls204`) THEN 1
                    ELSE 0
            END
            ) AS is_1_pf,
            (
            CASE 
                    WHEN appln_id_2 NOT IN(
                                    SELECT DISTINCT appln_id
                                    FROM `usptobias.patstat.tls204`) THEN 1
                    ELSE 0
            END
            ) AS is_2_pf
    FROM
            `usptobias.results_docdb.USExactTwins`
    """

# Defining the query job
query_job = client.query(sql, location='US', job_config=job_config)

query_job.result()

# 4.2. Getting the resulting table

dataset_id = 'results_docdb'
table_id = 'backbone_0'
# Getting the table's information
table_ref = client.dataset(dataset_id).table(table_id)
table = client.get_table(table_ref)

# Print tables information
print(table.schema)
print(table.description)
print("\nThere are {:,} number of records in total!".format(table.num_rows))

# Loading the first 10,000 rows
dataset_id = 'results_docdb'
table_id = 'backbone_0'
# Getting the table object
table_ref = client.dataset(dataset_id).table(table_id)
table = client.get_table(table_ref)

# Load the first 10000 rows into RowIterator object
rows = client.list_rows(table, max_results=10000)

# Loading the RowIterator object into Pandas dataframe
df = rows.to_dataframe()

df.head()

# 4.3. Exporting the table

# from google.cloud import bigquery
client = bigquery.Client()
bucket_name = 'patent-results'
folder_name = 'backbone'

project = 'usptobias'
dataset_id = 'results_docdb'
table_id = 'backbone_0'

destination_uri = 'gs://{}/{}/{}'.format(bucket_name, folder_name, 'backbone_0_docdb_*.csv.gz')
dataset_ref = client.dataset(dataset_id, project=project)
table_ref = dataset_ref.table(table_id)

job_config = bigquery.job.ExtractJobConfig()

job_config.compression = "GZIP"
job_config.destinationFormat = "CSV"

extract_job = client.extract_table(
    table_ref,
    destination_uri,
    # Location must match that of the source table.
    location='US',
    job_config=job_config)  # API request
extract_job.result()  # Waits for job to complete.
7
print('Exported {}:{}.{} to {}'.format(
    project, dataset_id, table_id, destination_uri))

# 5. Selecting Only "Exact Twins"

client = bigquery.Client()

## 5.1. Query to generate the table

job_config = bigquery.QueryJobConfig()
job_config.user_query_cache = False
# Set configuration.query.createDisposition
job_config.create_disposition = 'CREATE_IF_NEEDED'
# Set configuration.query.writeDisposition
job_config.write_disposition = 'WRITE_TRUNCATE' #'WRITE_APPEND'


# Set destination table
dataset_id = 'results_docdb'
table_id = 'USOnlyTwins'
table_ref = client.dataset(dataset_id).table(table_id)
job_config.destination = table_ref

sql= """
    SELECT
            * EXCEPT (exact_twins, docdb_nb_pairs)
    FROM
            `usptobias.results_docdb.USExactTwins`
    WHERE
            exact_twins=1
            AND appln_kind_2='A '
    """

# Defining the query job
query_job = client.query(sql, location='US', job_config=job_config)

query_job.result()

## 5.2. Getting the resulting table info

dataset_id = 'results_docdb'
table_id = 'USOnlyTwins'
# Getting the table's information
table_ref = client.dataset(dataset_id).table(table_id)
table = client.get_table(table_ref)

# Print tables information
print(table.schema)
print(table.description)
print("\nThere are {:,} number of records in total!".format(table.num_rows))

## 5.3. Loading the First 10,000 rows

dataset_id = 'results_docdb'
table_id = 'USOnlyTwins'
# Getting the table object
table_ref = client.dataset(dataset_id).table(table_id)
table = client.get_table(table_ref)

# Load the first 10000 rows into RowIterator object
rows = client.list_rows(table, max_results=10000)

# Loading the RowIterator object into Pandas dataframe
df = rows.to_dataframe()

df.head(7)

## 5.4.Deleting the table

# from google.cloud import bigquery
# client = bigquery.Client()
dataset_id = 'results_docdb'
table_id = 'USOnlyTwins'

table_ref = client.dataset(dataset_id).table(table_id)
#client.delete_table(table_ref)  # API request

print("Table {}:{} deleted.".format(dataset_id, table_id))

## 5.5. Exporting Table to GCP Bucket

# from google.cloud import bigquery
client = bigquery.Client()
bucket_name = 'patent-results'
folder_name = 'exactTwins'

project = 'usptobias'
dataset_id = 'results_docdb'
table_id = 'USExactTwins'

destination_uri = 'gs://{}/{}/{}'.format(bucket_name, folder_name, 'exact_twins_docdb-*.csv.gz')
dataset_ref = client.dataset(dataset_id, project=project)
table_ref = dataset_ref.table(table_id)

job_config = bigquery.job.ExtractJobConfig()

job_config.compression = "GZIP"
job_config.destinationFormat = "CSV"

extract_job = client.extract_table(
    table_ref,
    destination_uri,
    # Location must match that of the source table.
    location='US',
    job_config=job_config)  # API request
extract_job.result()  # Waits for job to complete.
7
print('Exported {}:{}.{} to {}'.format(
    project, dataset_id, table_id, destination_uri))

# 6.Filtering to get patents which filed in years between 2001 and 2017

## 6.1. Query

client = bigquery.Client()

job_config = bigquery.QueryJobConfig()
job_config.user_query_cache = False

# Set destination table
dataset_id = 'results_docdb'
table_id = 'USOnlyTwins_2017'
table_ref = client.dataset(dataset_id).table(table_id)
job_config.destination = table_ref

sql= """
    SELECT
            *
    FROM
            `usptobias.results_docdb.USOnlyTwins`
    WHERE
            appln_filing_year_1 <= 2017
            AND appln_filing_year_2 <= 2017
    """

# Defining the query job
query_job = client.query(sql, location='US', job_config=job_config)

query_job.result()

## 6.2. Getting the resulting table information

dataset_id = 'results_docdb'
table_id = 'USOnlyTwins_2017'
# Getting the table's information
table_ref = client.dataset(dataset_id).table(table_id)
table = client.get_table(table_ref)

# Print tables information
print(table.schema)
print(table.description)
print("\nThere are {:,} number of records in total!".format(table.num_rows))

## 6.3. Loading the first 10,000 rows

dataset_id = 'results_docdb'
table_id = 'USOnlyTwins_2017'
# Getting the table object
table_ref = client.dataset(dataset_id).table(table_id)
table = client.get_table(table_ref)

# Load the first 10000 rows into RowIterator object
rows = client.list_rows(table, max_results=10000)

# Loading the RowIterator object into Pandas dataframe
df = rows.to_dataframe()

df.head()

# 7. Querying For Comparing

## 7.1. Copying FamilyInfo table (df_twins_1)

'''# from google.cloud import bigquery
client = bigquery.Client()
# Defining source_table
source_dataset = client.dataset('results', project='usptobias')
source_table_ref = source_dataset.table('familyInfo')
# Defining target table
target_dataset = client.dataset('results_docdb', project='usptobias')
dest_table_ref = target_dataset.table('familyInfo')
# Initializing the Job
job = client.copy_table(
    source_table_ref,
    dest_table_ref,
    # Location must match that of the source and destination tables.
    location='US')  # API request

# Running the job
job.result()  # Waits for job to complete.'''

## 7.2. Getting the rows that are not mutual

client = bigquery.Client()

job_config = bigquery.QueryJobConfig()
job_config.user_query_cache = False

# Set destination table
dataset_id = 'results_docdb'
table_id = 'not_mutual'
table_ref = client.dataset(dataset_id).table(table_id)
job_config.destination = table_ref

sql = """
    SELECT 
            appln_id_1,
            appln_id_2,
            t1.is_in_1,
            t2.is_in_2,
            exact_twins_1,
            exact_twins_2
            
    FROM
            (
            SELECT
                    appln_id_1,
                    appln_id_2,
                    exact_twins AS exact_twins_1,
                    1 AS is_in_1
            FROM
                    `usptobias.results_docdb.familyInformation`
            ) AS t1
            
        FULL OUTER JOIN
            (
            SELECT
                    appln_id_1,
                    appln_id_2,
                    exact_twins AS exact_twins_2,
                    1 AS is_in_2
            FROM
                    `usptobias.results_docdb.USExactTwins`
            WHERE
                    appln_filing_year_1 <> 2018
                    AND appln_filing_year_2 <> 2018
            ) AS t2

        USING(appln_id_1, appln_id_2)
    WHERE
            (t1.is_in_1 IS NULL OR t2.is_in_2 IS NULL)
            
     """
# Defining the query
query_job = client.query(
    sql,
    # Location must match that of the dataset(s) referenced in the query.
    location='US',
    job_config=job_config)

query_job.result()

## 7.3. Getting the resulting table

dataset_id = 'results_docdb'
table_id = 'not_mutual'
# Getting the table's information
table_ref = client.dataset(dataset_id).table(table_id)
table = client.get_table(table_ref)

# Print tables information
print(table.schema)
print(table.description)
print("\nThere are {:,} number of records in total!".format(table.num_rows))

dataset_id = 'results_docdb'
table_id = 'not_mutual'
# Getting the table object
table_ref = client.dataset(dataset_id).table(table_id)
table = client.get_table(table_ref)

# Load the first 10000 rows into RowIterator object
rows = client.list_rows(table, max_results=10000)

# Loading the RowIterator object into Pandas dataframe
df = rows.to_dataframe()

df.head(7)

### 7.3.1. Deleting table

# from google.cloud import bigquery
# client = bigquery.Client()
dataset_id = 'results_docdb'
table_id = 'not_mutual_info'

table_ref = client.dataset(dataset_id).table(table_id)
#client.delete_table(table_ref)  # API request

print("Table {}:{} deleted.".format(dataset_id, table_id))

## 7.4. Adding the information for appln_id_1 and appln_id_2

client = bigquery.Client()

job_config = bigquery.QueryJobConfig()
job_config.user_query_cache = False

# Set destination table
dataset_id = 'results_docdb'
table_id = 'not_mutual_info'
table_ref = client.dataset(dataset_id).table(table_id)
job_config.destination = table_ref

sql = """
    SELECT 
            t1.*,
            t2.appln_auth_1,
            t2.appln_kind_1,
            t2.appln_filing_year_1,
            t3.appln_auth_2,
            t3.appln_kind_2,
            t3.appln_filing_year_2,
            t2.docdb_family_id,
            t2.docdb_family_size
            
    FROM
            `usptobias.results_docdb.not_mutual` AS t1
    LEFT JOIN
            (
            SELECT
                    appln_id,
                    appln_auth AS appln_auth_1,
                    appln_kind AS appln_kind_1,
                    appln_filing_year AS appln_filing_year_1,
                    docdb_family_id,
                    docdb_family_size
            FROM
                    `usptobias.patstat.tls201`
            ) AS t2
            ON t1.appln_id_1 = t2.appln_id
    LEFT JOIN
            (
            SELECT
                    appln_id,
                    appln_auth AS appln_auth_2,
                    appln_kind AS appln_kind_2,
                    appln_filing_year AS appln_filing_year_2
            FROM
                    `usptobias.patstat.tls201`
            ) AS t3
            ON t1.appln_id_2 = t3.appln_id
     """
# Defining the query
query_job = client.query(
    sql,
    # Location must match that of the dataset(s) referenced in the query.
    location='US',
    job_config=job_config)

query_job.result()

## 7.4. Getting the resulting table: "not_mutual_info"

dataset_id = 'results_docdb'
table_id = 'not_mutual_info'
# Getting the table object
table_ref = client.dataset(dataset_id).table(table_id)
table = client.get_table(table_ref)

# Load the first 10000 rows into RowIterator object
rows = client.list_rows(table, max_results=10000)

# Loading the RowIterator object into Pandas dataframe
df = rows.to_dataframe()

df.head(7)

## 7.5. Exporting the table to GCP Bucket

# from google.cloud import bigquery
client = bigquery.Client()
bucket_name = 'patent-results'
folder_name = 'exactTwins'

project = 'usptobias'
dataset_id = 'results_docdb'
table_id = 'not_mutual_info'

destination_uri = 'gs://{}/{}/{}'.format(bucket_name, folder_name, 'not_mutual.csv.gz')
dataset_ref = client.dataset(dataset_id, project=project)
table_ref = dataset_ref.table(table_id)


job_config = bigquery.job.ExtractJobConfig()

job_config.compression = "GZIP"
job_config.destinationFormat = "CSV"

extract_job = client.extract_table(
    table_ref,
    destination_uri,
    # Location must match that of the source table.
    location='US',
    job_config=job_config)  # API request
extract_job.result()  # Waits for job to complete.
7
print('Exported {}:{}.{} to {}'.format(
    project, dataset_id, table_id, destination_uri))
